+++ apstools/devices/__init__.py
@@ -40,11 +40,20 @@ from .axis_tuner import AxisTunerMixin
 
 from .description_mixin import EpicsDescriptionMixin
 
+from .dict_device_support import dict_device_factory
 from .dict_device_support import make_dict_device
 
 from .epics_scan_id_signal import EpicsScanIdSignal
 
 from .eurotherm_2216e import Eurotherm2216e
+
+from .flyer_motor_scaler import FlyerBase
+from .flyer_motor_scaler import ActionsFlyerBase
+from .flyer_motor_scaler import ScalerMotorFlyer
+from .flyer_motor_scaler import SignalValueStack
+from .flyer_motor_scaler import _SMFlyer_Step_1
+from .flyer_motor_scaler import _SMFlyer_Step_2
+from .flyer_motor_scaler import _SMFlyer_Step_3
 
 from .kohzu_monochromator import KohzuSeqCtl_Monochromator