+++ apstools/plans/__init__.py
@@ -2,6 +2,7 @@ from .alignment import lineup
 from .alignment import tune_axes
 from .alignment import TuneAxis
 from .alignment import TuneResults
+
 from .command_list import command_list_as_table
 from .command_list import CommandFileReadError
 from .command_list import execute_command_list
@@ -11,11 +12,16 @@ from .command_list import parse_text_command_file
 from .command_list import register_command_handler
 from .command_list import run_command_file
 from .command_list import summarize_command_file
+
 from .doc_run import addDeviceDataAsStream
 from .doc_run import documentation_run
+
 from .input_plan import request_input
+
 from .nscan_support import nscan
+
 from .run_blocking_function_plan import run_blocking_function
+
 from .sscan_support import sscan_1D
 
 # -----------------------------------------------------------------------------